+++ app/build.gradle
@@ -77,4 +77,5 @@ dependencies {
     annotationProcessor 'com.github.bumptech.glide:compiler:4.13.2'
 
     implementation "com.github.skydoves:landscapist-glide:1.5.3"
+
 }
