+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="hu.mczinke.payment_manager">
+    package="hu.mczinke.nasa_apod_viewer">
+
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -12,12 +13,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Paymentmanager"
+        android:theme="@style/Theme.ApodViewer"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="hu.mczinke.nasa_apod_viewer.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Paymentmanager">
+            android:theme="@style/Theme.ApodViewer">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
