+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/ui/components/BottomBarScreen.kt
@@ -1,0 +1,31 @@
+package hu.mczinke.nasa_apod_viewer.ui.components
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.ui.graphics.vector.ImageVector
+
+sealed class BottomBarScreen(
+    val route: String,
+    val title: String,
+    val icon: ImageVector
+) {
+    object Home : BottomBarScreen(
+        route = "home",
+        title = "Home",
+        icon = Icons.Default.Home
+    )
+
+    object Favorites : BottomBarScreen(
+        route = "favorites",
+        title = "Favorites",
+        icon = Icons.Default.Star
+    )
+
+    object Search : BottomBarScreen(
+        route = "search",
+        title = "Search",
+        icon = Icons.Default.Search
+    )
+}
