+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/ui/components/FavoritesScreen.kt
@@ -1,0 +1,9 @@
+package hu.mczinke.nasa_apod_viewer.ui.components
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun FavoritesScreen() {
+    Text(text = "Favorite")
+}
