+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/ui/preview_parameter_providers/ApodParameterProvider.kt
@@ -1,0 +1,17 @@
+package hu.mczinke.nasa_apod_viewer.ui.preview_parameter_providers
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import hu.mczinke.nasa_apod_viewer.models.Apod
+
+class ApodParameterProvider : PreviewParameterProvider<Apod> {
+    override val values = sequenceOf(
+        Apod(
+            "copyright",
+            "date",
+            "explanation",
+            "HDUrl",
+            "title",
+            "url",
+        )
+    )
+}
