+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/ui/preview_parameter_providers/StringParameterProvider.kt
@@ -1,0 +1,7 @@
+package hu.mczinke.nasa_apod_viewer.ui.preview_parameter_providers
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+
+class StringParameterProvider : PreviewParameterProvider<String> {
+    override val values = sequenceOf("string mock")
+}
