+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/ui/theme/Color.kt
@@ -1,0 +1,11 @@
+package hu.mczinke.nasa_apod_viewer.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val SpaceBlack = Color(0xFF19181B)
+val SpaceBlackVariant = Color(0xFF252525)
+val SpacePrimary = Color(0xFF2C5484)
+val SpacePrimaryVariant = Color(0xFF80A7C4)
+val SpaceSecondary = Color(0xFFCDCBDA)
+
+val DimmedWhite = Color(0xFFF7F7F7)
