+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/viewmodels/MainViewModel.kt
@@ -1,0 +1,24 @@
+package hu.mczinke.nasa_apod_viewer.viewmodels
+
+import android.util.Log
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import hu.mczinke.nasa_apod_viewer.BuildConfig
+import hu.mczinke.nasa_apod_viewer.models.Apod
+import kotlinx.coroutines.launch
+
+class MainViewModel(private val repository: Repository) : ViewModel() {
+
+    private val _apod: MutableLiveData<Apod> = MutableLiveData()
+    val apod: LiveData<Apod> = _apod
+
+    fun getAPOD() {
+        viewModelScope.launch {
+            val result = repository.getApod(BuildConfig.NASA_API_KEY)
+            Log.d("Response", result.toString())
+            _apod.value = result.toApod()
+        }
+    }
+}
