+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/viewmodels/MainViewModelFactory.kt
@@ -1,0 +1,12 @@
+package hu.mczinke.nasa_apod_viewer.viewmodels
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class MainViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        return MainViewModel(repository) as T
+    }
+}
