+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/viewmodels/SearchViewModelFactory.kt
@@ -1,0 +1,13 @@
+package hu.mczinke.nasa_apod_viewer.viewmodels
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+
+class SearchViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        return SearchViewModel(repository) as T
+    }
+}
