--- app/src/main/java/hu/mczinke/payment_manager/ui/components/FavoritesScreen.kt
@@ -1,9 +1,0 @@
-package hu.mczinke.payment_manager.ui.components
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun FavoritesScreen() {
-    Text(text = "Favorite")
-}
