--- app/src/main/java/hu/mczinke/payment_manager/ui/preview_parameter_providers/StringParameterProvider.kt
@@ -1,7 +1,0 @@
-package hu.mczinke.payment_manager.ui.preview_parameter_providers
-
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-
-class StringParameterProvider : PreviewParameterProvider<String> {
-    override val values = sequenceOf("string mock")
-}
