--- app/src/main/java/hu/mczinke/payment_manager/viewmodels/MainViewModelFactory.kt
@@ -1,12 +1,0 @@
-package hu.mczinke.payment_manager.viewmodels
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-class MainViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return MainViewModel(repository) as T
-    }
-}
