--- app/src/main/java/hu/mczinke/payment_manager/viewmodels/RetrofitInstance.kt
@@ -1,19 +1,0 @@
-package hu.mczinke.payment_manager.viewmodels
-
-import hu.mczinke.payment_manager.nasa_api.ApodApi
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-
-    private val retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://api.nasa.gov/planetary/apod/")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    val api: ApodApi by lazy {
-        retrofit.create(ApodApi::class.java)
-    }
-}
