--- app/src/main/java/hu/mczinke/payment_manager/viewmodels/SearchViewModelFactory.kt
@@ -1,13 +1,0 @@
-package hu.mczinke.payment_manager.viewmodels
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-
-class SearchViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return SearchViewModel(repository) as T
-    }
-}
