+++ app/src/main/res/values/themes.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.Paymentmanager" parent="android:Theme.Material.Light.NoActionBar">
+    <style name="Theme.ApodViewer" parent="android:Theme.Material.Light.NoActionBar">
         <item name="android:statusBarColor">@color/space_primary</item>
     </style>
 </resources>
